Create level-001
output:

i hope you didnt translate it by hand. thats what computers are for.doing it in by hand is inefficient and that’s why this text is so long.using string.maketrans() is recommended. now apply on the url.
ocr

use 'ocr' to replace 'map' in the url and you get into the next level.